create choose seat view

+++ lib/core/widgets/custom_button.dart
@@ -8,7 +8,7 @@ class CustomButton extends StatelessWidget {
       {super.key, required this.text, this.onPressed, this.height = 60});
   final String text;
   final void Function()? onPressed;
-  final int? height;
+  final double? height;
   @override
   Widget build(BuildContext context) {
     return MaterialButton(
